requirements.txt
anyio==3.5.0
asgiref==3.5.0
boto3==1.21.18
botocore==1.24.18
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.75.0
FastAPI-SQLAlchemy==0.2.1
ffmpeg-python==0.2.0
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jmespath==0.10.0
psycopg2-binary==2.9.3
pydantic==1.9.0
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
